+++ tests/common/include/tests/common/io_read_detector.inl
@@ -15,7 +15,7 @@
 #include "detray/core/detector.hpp"
 #include "detray/core/transform_store.hpp"
 #include "detray/io/csv_io.hpp"
-#include "tests/common/read_geometry.hpp"
+#include "tests/common/tools/read_geometry.hpp"
 
 /// @note __plugin has to be defined with a preprocessor command
