+++ tests/common/include/tests/common/tools/ray_gun.hpp
